requirements.txt
streamlit
langchain
openai
python-docx
tiktoken
faiss-cpu
python-dotenv
